Refactor gen 9 evolutions
